♻️  smol,tokio: Split `notified` module into a hierarchy
