+++ zlink-smol/src/notified/mod.rs
@@ -1,0 +1,7 @@
+//! Convenience API for maintaining state, that notifies on changes.
+
+mod once;
+mod state;
+
+pub use once::{Once, Stream as OnceStream};
+pub use state::{State, Stream};
